+++ packages/cloudflare/src/types/CloudflareConfig.ts
@@ -1,6 +1,9 @@
+import { CustomDomainsAdapter } from "@authhero/adapter-interfaces";
+
 export interface CloudflareConfig {
   zoneId: string;
   authKey: string;
   authEmail: string;
   enterprise?: boolean;
+  customDomainAdapter: CustomDomainsAdapter;
 }