+++ packages/cloudflare/test/helpers/test-server.ts
@@ -1,0 +1,27 @@
+import { Kysely, SqliteDialect } from "kysely";
+import SQLite from "better-sqlite3";
+import createAdapters, {
+  Database,
+  migrateToLatest,
+} from "@authhero/kysely-adapter";
+
+type getEnvParams = {
+  testTenantLanguage?: string;
+  emailValidation?: "enabled" | "enforced" | "disabled";
+};
+
+export async function getTestServer(args: getEnvParams = {}) {
+  const dialect = new SqliteDialect({
+    database: new SQLite(":memory:"),
+  });
+
+  // Don't use getDb here as it will reuse the connection
+  const db = new Kysely<Database>({ dialect: dialect });
+
+  await migrateToLatest(db, false);
+
+  return {
+    data: createAdapters(db),
+    db,
+  };
+}